requirements.txt
Django==5.1.3
django-unfold==0.40.0
